feat(io): composable port directions and runtime connection checking (#231)
* refactor(io)!: remove Undirected from type system

Remove the `Undirected` marker trait, and make direction-setting
wrappers like `Input<T>` act as an override to the direction of their
contained type (instead of direction being a property that can only be
set once).

This allows for use cases such as snooping on a bus, where the
(directed) bus type is put inside `Input<_>` to convert all signals into
inputs.

Due to coherence rules, the transitive `From` implementation for the
wrappers had to be dropped (previously, if `T: From<U>`, then `Input<T>:
From<U>`). This is because a user could implement `T: From<Input<T>>`,
which would cause our blanket implementation to apply for `Input<T>:
From<Input<T>>`, conflicting with std's blanket impl.

* refactor(io)!: remove direction type wrappers from instantiated type

Since there is no way to propagate direction down the hierarchy (so that
e.g. `foo.bar` has the correct direction at type level when `foo` is
wrapped in Input), remove direction information completely from the type
level on instantiated data types.

* feat(schematic): runtime driver checking of connected nets

Implement runtime tracking of driver types (directions) of all nodes
connected in a net, and validate that connected nodes are legal.

Error reporting is currently a basic stub implementation (panic at the
call where a bad connection was made), this should be improved later.

* refactor(io): pull out type -> schematic data instantation code

Factor out code to instantiate a SchematicType::Data inside a cell
builder to some common functions.

* fix(io): make driver conflict a recoverable warning

Ensure that the connection is still made even if there is a driver issue
and change the temporary error handling to emit a warning instead of
panicking.

* impl display for direction

---------

Co-authored-by: Rahul Kumar <rahulkumar@berkeley.edu>